fully working example with rgb/cmy/w simulation

+++ stm32-board/src/led_config.c
@@ -3,4 +3,4 @@
 
 uint16_t frame_buffer[FRAME_BUFFER_LENGTH];
 
-uint8_t current_usb_frame = 0;
+uint16_t current_usb_frame = 0;
